// client/src/api/generated/models/updateWebhookDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Knowted API
 * The Knowted API documentation
 * OpenAPI spec version: 1.0
 */

export interface UpdateWebhookDto {
  /** Webhook name */
  name?: string;
  /** Webhook URL */
  url?: string;
  /** Whether webhook is active */
  is_active?: boolean;
}